BotClientWrapper to entrypoint
